Build uberjar with lein uberjar at command line
I've switched out the build system to use lein because it easier to
use it to build uberjars. To run your repl you will need to use the
clojure lein/nrepl version in calva. The jar does get built but it
might need some more work to get it running.